adding Lambda layers
Adding layers into the project to simplify dependencies across functions and speed up build/deploys
