--- src/functions/titles/api_titles_create/requirements.txt
@@ -1,1 +1,0 @@
-jsonschema==2.6.0
