+++ src/functions/titles/api_titles_delete/api_titles_delete.py
@@ -1,5 +1,9 @@
 import logging
 import os
+import sys
+
+# Add '/opt' to the PATH for JamfPackages Lambda Layer
+sys.path.append('/opt')
 
 # from aws_xray_sdk.core import xray_recorder
 # from aws_xray_sdk.core import patch
