--- src/functions/titles/api_titles_delete/requirements.txt
@@ -1,1 +1,0 @@
-git+https://github.com/brysontyrrell/Opossum.git#egg=opossum
