+++ src/functions/titles/api_titles_update/api_titles_update.py
@@ -3,12 +3,18 @@ import io
 import json
 import logging
 import os
+import sys
+
+# Add '/opt' to the PATH for JamfPackages Lambda Layer
+sys.path.append('/opt')
 
 # from aws_xray_sdk.core import xray_recorder
 # from aws_xray_sdk.core import patch
 import boto3
 from botocore.exceptions import ClientError
 from jsonschema import validate, ValidationError
+
+from api_helpers import response
 
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
@@ -24,28 +30,6 @@ s3_bucket = boto3.resource('s3').Bucket(os.getenv('TITLES_BUCKET'))
 
 with open('schemas/schema_version.json', 'r') as f_obj:
     schema_version = json.load(f_obj)
-
-
-def response(message, status_code):
-    """Returns a dictionary object for an API Gateway Lambda integration
-    response.
-
-    :param message: Message for JSON body of response
-    :type message: str or dict
-
-    :param int status_code: HTTP status code of response
-
-    :rtype: dict
-    """
-    if isinstance(message, str):
-        message = {'message': message}
-
-    return {
-        'isBase64Encoded': False,
-        'statusCode': status_code,
-        'body': json.dumps(message),
-        'headers': {'Content-Type': 'application/json'}
-    }
 
 
 def read_table_entry(title_id, contributor_id):
