+++ src/functions/titles/api_titles_update/requirements.txt
@@ -1,1 +1,1 @@
-jsonschema==2.6.0
+
