+++ src/layers/api_shared/api_helpers.py
@@ -1,0 +1,23 @@
+import json
+
+
+def response(message, status_code):
+    """Returns a dictionary object for an API Gateway Lambda integration
+    response.
+
+    :param message: Message for JSON body of response
+    :type message: str or dict
+
+    :param int status_code: HTTP status code of response
+
+    :rtype: dict
+    """
+    if isinstance(message, str):
+        message = {'message': message}
+
+    return {
+        'isBase64Encoded': False,
+        'statusCode': status_code,
+        'body': json.dumps(message),
+        'headers': {'Content-Type': 'application/json'}
+    }
